requirements.txt
streamlit==1.32.0
cryptography==42.0.2
pycryptodome==3.20.0
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6